Feat(renderer)!: Changed color format

+++ src/renderer.hpp
@@ -76,8 +76,8 @@ typedef struct {
 } mosaic_control;
 
 typedef struct {
-    float red;
-    float green;
-    float blue;
-    float alpha;
+    uint8_t red;
+    uint8_t green;
+    uint8_t blue;
+    uint8_t alpha;
 } color;
